refactor(analysis-types): split effort DTOs
